pyproject.toml
[tool.poetry]
name = "app-service"
version = "0.1.0"
description = ""
authors = ["Tiger Analytics <tiger@tigeranalytics.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
SQLAlchemy = "^1.4.41"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "^0.21.0"
manage-fastapi = "^1.1.1"
sqlmodel = "^0.0.8"
mysql-connector-python = "^8.1.0"
pyodbc = "4.0.35"
azure-identity = "^1.11.0"
python-jose = "^3.3.0"
pylint = "^2.15.6"
bcrypt = "^4.0.1"
aioodbc = "^0.3.3"
numpy = "^1.23.5"
shapely = "^1.8.5.post1"
pandas = "^1.5.2"
passlib="1.7.4"
python-multipart="0.0.5"
loguru = "^0.6.0"
openpyxl = "^3.0.10"
ortools = "^9.5.2237"
matplotlib = "^3.7.1"
geopy = "^2.3.0"
cachetools = "^5.3.0"
httpx = "^0.23.3"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
autoflake = "^1.4"
flake8 = "^3.8.4"
mypy = "^0.991"
isort = "^5.0"
pre-commit = "^2.8.2"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

# pyproject.toml

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

